requirements.txt
# Additional requirements go here
djangorestframework==3.12.4
django-mptt==0.13.4
